+++ src/blueprint/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.3.2
++++++
+* Migrate to atomic commands
 
 0.3.1
 +++++